+++ common/font.h
@@ -57,14 +57,14 @@ class GraphicViewPortClass;
 void* Set_Font(void const* fontptr);
 int Char_Pixel_Width(char chr);
 unsigned int String_Pixel_Width(char const* string);
-void Get_Next_Text_Print_XY(GraphicViewPortClass& vp, unsigned long offset, int* x, int* y);
+void Get_Next_Text_Print_XY(GraphicViewPortClass& vp, unsigned int offset, int* x, int* y);
 void Set_Font_Palette_Range(void const* palette, int start_idx, int end_idx);
 void* Load_Font(char const* name);
 
 /*=========================================================================*/
 /* The following prototypes are for the file: TEXTPRNT.ASM                 */
 /*=========================================================================*/
-long Buffer_Print(void* thisptr, const char* str, int x, int y, int fcolor, int bcolor);
+int Buffer_Print(void* thisptr, const char* str, int x, int y, int fcolor, int bcolor);
 void* Get_Font_Palette_Ptr();
 /*=========================================================================*/
