+++ common/memflag.h
@@ -63,10 +63,10 @@ void* operator new(size_t size, MemoryFlagType flag);
 void* operator new[](size_t size, MemoryFlagType flag);
 void* Alloc(size_t bytes_to_alloc, MemoryFlagType flags);
 void Free(void const* pointer);
-void* Resize_Alloc(void* original_ptr, unsigned long new_size_in_bytes);
-long Ram_Free(MemoryFlagType flag);
-long Heap_Size(MemoryFlagType flag);
-long Total_Ram_Free(MemoryFlagType flag);
+void* Resize_Alloc(void* original_ptr, size_t new_size_in_bytes);
+int Ram_Free(MemoryFlagType flag);
+int Heap_Size(MemoryFlagType flag);
+int Total_Ram_Free(MemoryFlagType flag);
 
 //#pragma option -Jgd
 
@@ -85,9 +85,9 @@ inline void* operator new[](size_t size, MemoryFlagType flag)
 /* The following prototypes are for the file: MEM_COPY.ASM						*/
 /*=========================================================================*/
 
-void Mem_Copy(void const* source, void* dest, unsigned long bytes_to_copy);
+void Mem_Copy(void const* source, void* dest, unsigned int bytes_to_copy);
 
-inline void* Add_Long_To_Pointer(void const* ptr, long size)
+inline void* Add_Long_To_Pointer(void const* ptr, int size)
 {
     return ((void*)((char const*)ptr + size));
 }
@@ -95,7 +95,7 @@ inline void* Add_Long_To_Pointer(void const* ptr, long size)
 extern void (*Memory_Error)(void);
 extern void (*Memory_Error_Exit)(char* string);
 
-extern unsigned long MinRam; // Record of least memory at worst case.
-extern unsigned long MaxRam; // Record of total allocated at worst case.
+extern size_t MinRam; // Record of least memory at worst case.
+extern size_t MaxRam; // Record of total allocated at worst case.
 
 #endif
